+++ components/Footer.tsx
@@ -1,5 +1,5 @@
-import { Box, Link, Text } from "@chakra-ui/react";
-import Image from "next/image";
+import { Box, Link, Text } from "@chakra-ui/react"
+import Image from "next/image"
 
 export default function Footer () {
   return (
