+++ pages/_app.tsx
@@ -1,9 +1,9 @@
 import type { AppProps } from 'next/app'
 import { ChakraProvider } from '@chakra-ui/react'
 import '../styles/globals.css'
-import Navbar from "../components/Navbar";
-import Footer from '../components/Footer';
-import { reducer, StateProvider } from '../state';
+import Navbar from "../components/Navbar"
+import Footer from '../components/Footer'
+import { reducer, StateProvider } from '../state'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
