+++ pages/api/getWeatherData.ts
@@ -1,18 +1,9 @@
 import axios from "axios"
-import { NextApiRequest, NextApiResponse } from "next";
-import { AirQualityIndexFromApi } from "../../types/AirQualityIndex"
-import { CurrentWeatherFromApi } from "../../types/CurrentWeather"
-import { WeatherForecastFromApi } from "../../types/WeatherForecast"
-import { toAirQualityIndex, toCurrentWeather, toWeatherForecast } from "../../utils";
+import { NextApiResponse } from "next"
+import { AirQualityIndexFromApi, CurrentWeatherFromApi, WeatherApiRequest, WeatherForecastFromApi } from "../../types"
+import { toAirQualityIndex, toCurrentWeather, toWeatherForecast } from "../../utils"
 
-interface GetWeatherDataRequest extends NextApiRequest {
-  query: {
-    "lat": string;
-    "lon": string;
-  }
-}
-
-export default async function getWeatherData(req: GetWeatherDataRequest, res: NextApiResponse) {
+export default async function getWeatherData(req: WeatherApiRequest, res: NextApiResponse) {
   const BASE_URL = 'https://api.openweathermap.org/data/2.5'
   const appid = process.env.REACT_APP_API_KEY
   try {
@@ -21,13 +12,13 @@ export default async function getWeatherData(req: GetWeatherDataRequest, res: Ne
       lon: req.query.lon,
       appid,
     }
-  
+
     const weatherParams = {
       ...aqiParams,
       units: 'metric',
       lang: 'id',
     }
-  
+
     const [
       { data: weatherFromApi },
       { data: forecastFromApi },
@@ -37,14 +28,14 @@ export default async function getWeatherData(req: GetWeatherDataRequest, res: Ne
       axios.get<WeatherForecastFromApi>(`${BASE_URL}/forecast`, { params: weatherParams }),
       axios.get<AirQualityIndexFromApi>(`${BASE_URL}/air_pollution`, { params: aqiParams }),
     ])
-  
+
     res.json({
       currentWeather: toCurrentWeather(weatherFromApi),
       weatherForecast: toWeatherForecast(forecastFromApi),
       airQualityIndex: toAirQualityIndex(aqiFromApi),
     })
   } catch (error) {
-    res.status(500).json({error})
+    res.status(500).json({ error })
   }
-  
+
 }
