+++ utils/constant.ts
@@ -1,0 +1,2 @@
+export const BASE_URL = 'https://api.openweathermap.org/data/2.5'
+export const appid = process.env.REACT_APP_API_KEY
